added User Registration and password encryption

+++ admin/includes/view_all_posts.php
@@ -97,7 +97,7 @@ if(isset($_POST["checkBoxArray"])){
             echo "<td>{$post_comment_count}</td>";
             echo "<td>{$post_date}</td>";
             echo "<td><a href='posts.php?source=edit_post&ep={$post_id}'>Edit</a></td>";
-            echo "<td><a href='posts.php?delete={$post_id}'>Delete</a></td>";
+            echo "<td><a onClick=\" javascript: return confirm('Are you sure you want to delete this?'); \" href='posts.php?delete={$post_id}'>Delete</a></td>";
             echo "</tr>";
             }
